requirements.txt
requests
beautifulsoup4
pillow
sklearn
tkinter
lxml
